+++ instrumentation/opentelemetry-instrumentation-pika/tests/test_pika_instrumentation.py
@@ -1,0 +1,104 @@
+# Copyright The OpenTelemetry Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from unittest import TestCase, mock
+
+from pika.adapters import BaseConnection, BlockingConnection
+from pika.channel import Channel
+from wrapt import BoundFunctionWrapper
+
+from opentelemetry.instrumentation.pika import PikaInstrumentor
+from opentelemetry.trace import Tracer
+
+
+class TestPika(TestCase):
+    def setUp(self) -> None:
+        self.channel = mock.MagicMock(spec=Channel)
+        self.channel._impl = mock.MagicMock(spec=BaseConnection)
+        self.mock_callback = mock.MagicMock()
+        self.channel._impl._consumers = {"mock_key": self.mock_callback}
+
+    def test_instrument_api(self) -> None:
+        instrumentation = PikaInstrumentor()
+        instrumentation.instrument()
+        self.assertTrue(
+            isinstance(BlockingConnection.channel, BoundFunctionWrapper)
+        )
+        assert hasattr(
+            instrumentation, "__opentelemetry_tracer_provider"
+        ), "Tracer not stored for the object!"
+        instrumentation.uninstrument(channel=self.channel)
+        self.assertFalse(
+            isinstance(BlockingConnection.channel, BoundFunctionWrapper)
+        )
+
+    @mock.patch(
+        "opentelemetry.instrumentation.pika.PikaInstrumentor._instrument_channel_functions"
+    )
+    @mock.patch(
+        "opentelemetry.instrumentation.pika.PikaInstrumentor._instrument_consumers"
+    )
+    def test_instrument(
+        self,
+        instrument_consumers: mock.MagicMock,
+        instrument_channel_functions: mock.MagicMock,
+    ):
+        PikaInstrumentor.instrument_channel(channel=self.channel)
+        assert hasattr(
+            self.channel, "_is_instrumented_by_opentelemetry"
+        ), "channel is not marked as instrumented!"
+        instrument_consumers.assert_called_once()
+        instrument_channel_functions.assert_called_once()
+
+    @mock.patch("opentelemetry.instrumentation.pika.utils._decorate_callback")
+    def test_instrument_consumers(
+        self, decorate_callback: mock.MagicMock
+    ) -> None:
+        tracer = mock.MagicMock(spec=Tracer)
+        expected_decoration_calls = [
+            mock.call(value, tracer, key)
+            for key, value in self.channel._impl._consumers.items()
+        ]
+        PikaInstrumentor._instrument_consumers(
+            self.channel._impl._consumers, tracer
+        )
+        decorate_callback.assert_has_calls(
+            calls=expected_decoration_calls, any_order=True
+        )
+        assert all(
+            hasattr(callback, "_original_callback")
+            for callback in self.channel._impl._consumers.values()
+        )
+
+    @mock.patch(
+        "opentelemetry.instrumentation.pika.utils._decorate_basic_publish"
+    )
+    def test_instrument_basic_publish(
+        self, decorate_basic_publish: mock.MagicMock
+    ) -> None:
+        tracer = mock.MagicMock(spec=Tracer)
+        original_function = self.channel.basic_publish
+        PikaInstrumentor._instrument_basic_publish(self.channel, tracer)
+        decorate_basic_publish.assert_called_once_with(
+            original_function, self.channel, tracer
+        )
+        self.assertEqual(
+            self.channel.basic_publish, decorate_basic_publish.return_value
+        )
+
+    def test_uninstrument_channel_functions(self) -> None:
+        original_function = self.channel.basic_publish
+        self.channel.basic_publish = mock.MagicMock()
+        self.channel.basic_publish._original_function = original_function
+        PikaInstrumentor._uninstrument_channel_functions(self.channel)
+        self.assertEqual(self.channel.basic_publish, original_function)